backend/requirements.txt
# Web Framework & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database
motor>=3.6.0
pymongo>=4.9.0

# LangChain & AI
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.50.0

# Vector Store
faiss-cpu>=1.8.0
tiktoken>=0.7.0

# File Processing
pypdf>=4.3.0
python-docx>=1.1.2
openpyxl>=3.1.5